ocr/requirements.txt
fastapi
uvicorn
pytesseract
pillow
pythainlp
pdf2image
PyMuPDF
requests
fuzzywuzzy
os
python-dotenv
